requirements.txt
# ── Core AI / RL ─────────────────────────────────────────────
stable-baselines3[extra]>=2.3.0
torch>=2.2.0
gymnasium>=0.29.0

# ── Data ─────────────────────────────────────────────────────
numpy>=1.26.0
pandas>=2.2.0
scipy>=1.13.0

# ── Angel One API ────────────────────────────────────────────
smartapi-python>=1.3.5
pyotp>=2.9.0
websocket-client>=1.7.0

# ── Async / networking ───────────────────────────────────────
aiohttp>=3.9.0
aiosqlite>=0.20.0

# ── Utilities ────────────────────────────────────────────────
python-dotenv>=1.0.0
loguru>=0.7.2

# ── Dev / testing ────────────────────────────────────────────
pytest>=8.0.0
pytest-asyncio>=0.23.0
